MarketInstrumentRespone Total type fixed (from decimal to int)

+++ Insight.Tinkoff.Invest/Dto/Market/Payloads/MarketInstrumentResponsePayload.cs
@@ -4,7 +4,7 @@ namespace Insight.Tinkoff.Invest.Dto.Payloads
 {
     public sealed class MarketInstrumentResponsePayload 
     {
-        public decimal Total { get; set; }
+        public int Total { get; set; }
 
         public List<MarketInstrument> Instruments { get; set; }
     }
